feat: Implement String  concatenation with test
