+++ include/efp.hpp
@@ -11,6 +11,7 @@
 #include "c_utility.hpp"
 #include "sm.hpp"
 #include "string.hpp"
+#include "format.hpp"
 #include "io.hpp"
 
 #endif
