+++ test/test.cpp
@@ -9,3 +9,4 @@
 #include "c_utility_test.hpp"
 #include "sm_test.hpp"
 #include "string_test.hpp"
+#include "format_test.hpp"
